requirements.txt
ipykernel
streamlit
google-generativeai
python-dotenv